pyproject.toml
[project]
name = "folio-api"
version = "0.1.0"
description = "API for FOLIO - the Federated Open Legal Information Ontology"
authors = [{ name = "ALEA Institute", email = "hello@aleainstitute.ai" }]
requires-python = ">=3.10,<4.0.0"
readme = "README.md"
license = "MIT"
dependencies = [
    "folio-python[search]>=0.1.5",
    "fastapi>=0.112.2",
    "uvicorn>=0.30.6",
]

[project.urls]
Homepage = "https://folio.openlegalstandard.org"
Repository = "https://github.com/alea-institute/folio-api"

[dependency-groups]
dev = [
    "sphinx>=7.4.7,<8",
    "black>=24.4.2,<25",
    "myst-parser>=3.0.1,<4",
    "sphinx-book-theme>=1.1.3,<2",
    "sphinxcontrib-mermaid>=0.9.2,<0.10",
    "pytest>=8.3.1,<9",
    "pytest-asyncio>=0.23.8,<0.24",
    "pytest-cov>=5.0.0,<6",
    "pylint>=3.2.7,<4",
]

[tool.hatch.build.targets.sdist]
include = ["folio_api"]

[tool.hatch.build.targets.wheel]
include = ["folio_api"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pylint]
max-line-length = 120
disable = [
    "line-too-long",
    "too-few-public-methods",
    "no-self-argument",
    "cyclic-import"
]
ignore-patterns = ["test_.*?.py", "conftest.py"]

[tool.pylint.messages_control]
max-args = 10
max-attributes = 15
max-branches = 15
max-locals = 25
min-public-methods = 0

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
addopts = "--cov=folio_api --cov-report=term-missing --cov-report=xml"
